+++ src/Pages/Accounts/Components/AccountTransactions/index.ts
@@ -1,2 +1,0 @@
-export * from './AccountTxsChart';
-export * from './AccountTxs';